Make the ROS node naming consistent.
Node executables are named accordingly to the package name with '_node'
appended. If multiple nodes belong to the same package, then a scheme
like e.g. caros_robotiq3_node and caros_robotiq2_node is used.
The default ROS namespace for ROS launch files are the name of the node
without the '_node'.