Add support for roslint on caros_common.
Also made changes according to roslint. Ignoring some roslint checks, as
they require stupid workarounds elsewhere (e.g. runtime/references which
would require a cast to non-const when using the ROS API, just to keep
roslint from complaining.). It could also be possible to manually
specify files to be checked by roslint, but having to keep that list up-to-date is not preferable.